docker-jobs/parallel-file-insert-blazegraph-job/requirements.txt
ray~=1.4.1
progressbar2
sparqlwrapper